+++ src/pages/UserGender/ui/UserGenderPage.style.ts
